refactor: handlers에서 mockDB분리로 백엔드단 로직분리
